fix: read user grants as grant owner

+++ internal/management/repository/eventsourcing/eventstore/user_grant.go
@@ -131,7 +131,7 @@ func (repo *UserGrantRepo) SearchUserGrants(ctx context.Context, request *grant_
 	result = &grant_model.UserGrantSearchResponse{
 		Offset:      request.Offset,
 		Limit:       request.Limit,
-		TotalResult: uint64(count),
+		TotalResult: count,
 		Result:      model.UserGrantsToModel(grants),
 	}
 	if sequenceErr == nil {
